TFTP config upload
Configuration file upload working

+++ extern.h
@@ -31,7 +31,7 @@ extern volatile int32_t*   ptrJointFeedback[JOINTS];
 extern volatile uint8_t*   ptrJointEnable;
 extern volatile float*     ptrSetPoint[VARIABLES];
 extern volatile float*     ptrProcessVariable[VARIABLES];
-extern volatile uint16_t*  ptrInputs;
-extern volatile uint16_t*  ptrOutputs;
+extern volatile uint32_t*  ptrInputs;
+extern volatile uint32_t*  ptrOutputs;
 
 #endif
